Final App
There are four different pages in this app, All Notes, Favorites, Reminders and Hidden
Page 1 - ‘All Notes’ page will show all the notes saved in the app except for the hidden notes. It will give options to add new notes, delete notes, add to favorites option, and hide your notes option.
Page 2: This page is known as ‘Favorites’. It will include only those notes that were added from “All Notes” page to the favorites.
Page 2: ‘Hidden’ page will require a password every time the user visits this page. Notes saved in this page will not be available in any of the other pages of the app.
Page 3: in this page, user can add notes separately in it just to set reminders to them, such as alarm. It will then give reminder to the user at the exact same date and time, the user set.